[LEMS-2737/move-scoring-logic] rename WidgetLogic to CoreWidgetRegistry
