+++ packages/perseus-core/src/index.ts
@@ -120,7 +120,7 @@ export {
 
 export type * from "./widgets/logic-export.types";
 
-export * as WidgetLogic from "./widgets/widget-registry";
+export * as CoreWidgetRegistry from "./widgets/core-widget-registry";
 
 export {default as getOrdererPublicWidgetOptions} from "./widgets/orderer/orderer-util";
 export {default as getCategorizerPublicWidgetOptions} from "./widgets/categorizer/categorizer-util";
