+++ packages/perseus-core/src/widgets/core-widget-registry.ts
