+++ packages/perseus-editor/src/diffs/renderer-diff.tsx
@@ -3,7 +3,7 @@
  */
 
 import {Widgets} from "@khanacademy/perseus";
-import {WidgetLogic} from "@khanacademy/perseus-core";
+import {CoreWidgetRegistry} from "@khanacademy/perseus-core";
 import * as React from "react";
 import _ from "underscore";
 
@@ -21,7 +21,7 @@ const filterWidgetInfo = function (widgetInfo, showAlignmentOptions: boolean) {
     // Show alignment options iff multiple valid ones exist for this widget
     if (
         showAlignmentOptions &&
-        WidgetLogic.getSupportedAlignments(type).length > 1
+        CoreWidgetRegistry.getSupportedAlignments(type).length > 1
     ) {
         // @ts-expect-error - TS2339 - Property 'alignment' does not exist on type '{ readonly options: any; }'.
         filteredWidgetInfo.alignment = alignment;
