+++ packages/perseus-editor/src/editor.tsx
@@ -6,7 +6,11 @@ import {
     Util,
     Widgets,
 } from "@khanacademy/perseus";
-import {Errors, PerseusError, WidgetLogic} from "@khanacademy/perseus-core";
+import {
+    CoreWidgetRegistry,
+    Errors,
+    PerseusError,
+} from "@khanacademy/perseus-core";
 import $ from "jquery";
 // eslint-disable-next-line import/no-unresolved, import/no-extraneous-dependencies
 import katex from "katex";
@@ -673,7 +677,8 @@ class Editor extends React.Component<Props, State> {
         const widgetContent = widgetPlaceholder.replace("{id}", id);
 
         // Add newlines before block-display widgets like graphs
-        const isBlock = WidgetLogic.getDefaultAlignment(widgetType) === "block";
+        const isBlock =
+            CoreWidgetRegistry.getDefaultAlignment(widgetType) === "block";
 
         const prelude = oldContent.slice(0, cursorRange[0]);
         const postlude = oldContent.slice(cursorRange[1]);
